move happy init/exec interop test under package msg
